+++ majordome/openfoam/__init__.py
@@ -7,4 +7,12 @@ __getattr__, __dir__ = setup_lazy_exports(__name__, globals(), {
     "FoamTabularData": ".post",
     "FoamLagrangianTable": ".post",
     "FoamPostProcessingLoader": ".post",
+}, linux_exports={
+    # run:
+    "FoamHelpers": ".run",
+    "FoamArguments": ".run",
+    "FoamCleaner": ".run",
+    "FoamRunner": ".run",
+    "FoamMeshing": ".run",
+    "FoamProject": ".run",
 })